requirements.txt
python-telegram-bot==20.3
supabase
requests
beautifulsoup4
playwright
flask
gunicorn
